backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
openai==1.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
mlflow==2.8.1
dvc==3.30.1
prometheus-client==0.17.1
python-gitlab==3.15.0
PyGithub==2.1.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2 